news/requirements.txt
feedparser
pandas
beautifulsoup4
lxml
websocket-client
PyYAML
pandas-gbq
psycopg2-binary
SQLAlchemy
holidays
yfinance
openai
pytz
twilio
spacy-transformers
scikit-learn
joblib